[MAT-132] fix: 순환참조 해결
